gestion_citas_medicas/requirements.txt
django
django-cors-headers
django-rest-framework
psycopg2